ui/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
pillow==10.4.0
aiofiles==24.1.0
pydantic==2.8.0
pytest==8.3.0
pytest-asyncio==0.23.8
respx==0.21.1
